backend/requirements.txt
fastapi==0.118.0
sqlalchemy==2.0.43
pydantic[email]==2.11.10
uvicorn[standard]==0.37.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20